backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
nltk==3.9.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-dotenv==1.2.1
requests==2.32.5
